separated the code into files
