+++ index.js
@@ -1,6 +1,8 @@
 /**
  * @format
  */
-import {registerRootComponent} from 'expo';
-import AppWrapper from './App';
-registerRootComponent(AppWrapper);
+import {
+  registerRootComponent
+} from 'expo';
+import App from './App';
+registerRootComponent(App);
